core/pyproject.toml
[project]
name = "meet-asbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.2",
    "ffmpeg>=1.4",
    "flask>=3.1.2",
    "google-cloud-storage>=3.6.0",
    "google-genai>=1.52.0",
    "gunicorn>=23.0.0",
    "pydub>=0.25.1",
    "resend>=2.19.0",
    "selenium>=4.38.0",
    "webdriver-manager>=4.0.2",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pytest>=9.0.2",
]
